==> 02-LIMMA.Rmd <==
# LIMMA


Here we gonna use the limma package to perform the normalization of the data and the differential expression annalysis.
Limma is a Bioconductor package which uses linear model for analysing experiments and the assessment of differential expression.

## loading and visualizing the data 
```{r}
data<-read.table("SWATH.example.data.csv", sep=",",header=T, row.names = 1)
DT::datatable(data)
boxplot(data, ylab="log2(Intensity)", xlab="samples", main = "distribution of Intensity") 
hist(as.numeric(unlist(data)), main = "Histogram of Intensity distribution",  xlab = "Intensity")
```

## Normalization


### design and contrast matrix

In order to perform the normalization with the limma package we need to define the design matrix generated with model.matrix(), which identify which sample belong to which condition.    


```{r}
#for this specific data :
exp.design <- data.frame(samples = colnames(data), condition = 1) ## first we define a matrix containing the experiment design information by associating a sample name with the condition name
exp.design$condition[1:10] = "neonates"
exp.design$condition[11:20] = "lessone"
exp.design$condition[21:30] = "onetofive"
exp.design$condition[31:40] = "adult"

design <- model.matrix(~0 + exp.design$condition, data = exp.design) ## model.matrix() use the experiment design to generate a matrix.
colnames(design) <- sort(unique(exp.design$condition))
row.names(design) <- exp.design$samples
DT::datatable(as.data.frame(design))
```


We also need to create a contrast matrix generated with the limma function makeContrast(). This matrix allows to define the comparisons between the differents conditions to perform.

```{r}
contrast <- makeContrasts(adult-neonates, lessone-neonates, onetofive-neonates, levels=design)
DT::datatable(as.data.frame(contrast))
```


### applying limma

Now, we can normalise the dataset using the following commands. The calcNormFactors(), calculates the normalization factors to scale the library sizes.

The limma package (since version 3.16.0) offers the voom function that will normalise the proteins intensity and apply a linear model to the normalised data before computing moderated t-statistics of differential expression.

The returned data object consists of a few attributes, which you can check using names(y), one of which is the normalised expression (y$E) values in log2 scale.

```{r}
library(limma)
library(edgeR)

dge <- DGEList(data)
dge <- calcNormFactors(dge)
y <- voom(dge, design)
norm.data <- y$E
DT::datatable(as.data.frame(norm.data))
boxplot(norm.data, ylab="normalized Intensity", xlab="samples", main = "distribution of the normalized Intensity")
hist(norm.data, main = "Histogram of Intensity distribution",  xlab = "Intensity")
```


## Differential expression annalysis
### fitting the model

To fit the model, use the lmFit() function, which takes in the normalised data object and the design matrix:

```{r}
fit <- lmFit(y, design)
```

Refit the model using the comparisons defined:
```{r}
fit2 <- contrasts.fit(fit, contrast)
fit2 <- eBayes(fit2)
```


### Extracting the results 

The topTable function summarises the output from limma in a table format.
```{r}
adult.vs.neonates <- topTable(fit2, coef = "adult - neonates", n = nrow(fit2))
```
```{r}
lessone.vs.neonates <- topTable(fit2, coef = "lessone - neonates", n = nrow(fit2))
```
```{r}
onetofive.vs.neonates <- topTable(fit2, coef = "onetofive - neonates", n = nrow(fit2))
```

```{r}
DT::datatable(adult.vs.neonates, caption = "adult vs neonates")
DT::datatable(lessone.vs.neonates, caption = "less than one vs neonates")
DT::datatable(onetofive.vs.neonates, caption = "one to five vs neonates")
```



We keep only the values of interest like the p value, the adjusted p value and the fold change to create the volcano plot. The results for the different pairwise comparison are stored in a list of dataframe. 
```{r}
adult.vs.neonates$protein <- row.names(adult.vs.neonates)
row.names(adult.vs.neonates) <- NULL
adult.vs.neonates <- adult.vs.neonates[, c(7, 4, 5, 1)]

lessone.vs.neonates$protein <- row.names(lessone.vs.neonates)
row.names(lessone.vs.neonates) <- NULL
lessone.vs.neonates <- lessone.vs.neonates[, c(7, 4, 5, 1)]

onetofive.vs.neonates$protein <- row.names(onetofive.vs.neonates)
row.names(onetofive.vs.neonates) <- NULL
onetofive.vs.neonates <- onetofive.vs.neonates[, c(7, 4, 5, 1)]

Results <- list(adult.vs.neonates, lessone.vs.neonates, onetofive.vs.neonates)
```




### Differentially expressed proteins

A protein is considered differentially expressed if its p.value and fold change are superior to a determined threshold. The fold change represent how much a protein is differentially expressed between two conditions and the p value allows to evaluate the statistical significance of this difference.

```{r}

thresh_fc <- 0.5
thresh_p <- 0.05
deProt <- Results
for(i in 1:length(Results)){
    fc = as.data.frame(deProt[[i]])[,4]
    p = as.data.frame(deProt[[i]])[,3]
    dt <-as.data.frame(deProt[[i]])
    deProt[[i]] <- dt[which(p<=thresh_p & abs(fc)>=thresh_fc),]
  }
```
```{r, echo=FALSE, warning=FALSE}
DT::datatable(as.data.frame(deProt[[1]]), caption = "adult vs neonates")
DT::datatable(as.data.frame(deProt[[2]]), caption = "less than one vs neonates")
DT::datatable(as.data.frame(deProt[[3]]), caption = "one to five vs neonates")
```



### Volcano Plot

For each pairwise comparison in the results list a volcano plot is created, volcano plots are the best way to visualise the p value and the fold change at the same time. 

```{r, warning=FALSE}

library(ggplot2)
library(plotly, quietly = TRUE) # using the plotly package to add interractivity to the plot
library(ggrepel)

tresh_fc = 0.5 ## Fold change threshold
tresh_p = 0.05 ## p.value treshold
nameVplot = c("adult vs neonates","less than one vs neonates","one to five vs neonates")
VP <- list() #list containing the volcano to display them after the loop


for(i in 1:length(Results)){
  
  values <- as.data.frame(Results[[i]])
  forplot <- data.frame(x=as.numeric(values[,4]), y=-log10(values[,3]), id=as.character(values[,1]))
  tmp <- forplot[as.numeric(forplot$y)>=-log10(tresh_fc) & abs(forplot$x)>tresh_fc,]
  p <- ggplot(forplot) + geom_point(aes(x, y ,label = id, color = ifelse(y>=-log10(tresh_p) & abs(x)>=tresh_fc,"DE proteins", "not significant")),show.legend = F) +
    scale_color_manual(values = c("red", "blue")) +
    geom_vline(xintercept = tresh_fc ) +
    geom_vline(xintercept = -tresh_fc) + 
    geom_hline(yintercept = -log10(tresh_p)) + 
    labs(title = nameVplot[i],x="log2(Fold-change)", y="-log10(P.Value)") + theme_bw() 
    VP[[i]] <- p
}
```

```{r, warning=FALSE}
layout(ggplotly(VP[[1]], tooltip = c("x","y","id")))
layout(ggplotly(VP[[2]], tooltip = c("x","y","id")))
layout(ggplotly(VP[[3]], tooltip = c("x","y","id")))
```


### Ven Diagramm

List of the proteins deemed as differentially expressed in all of the pairwise comparisons :

```{r}
res <- vector()
for(i in 1:length(deProt)){
  res <- append(res, as.character(deProt[[i]][,1]))
}
res <- unique(res)
saveRDS(res, "resLIMMA.rds") #saving the name of the differentially expressed proteins in a R object.
print(res)
```


This Venn diagram allows to visualize the number of proteins deemed to be differentially expressed in each pairwise comparison.


```{r}
library(VennDiagram)

grid.newpage()
draw.triple.venn(area1 = length(deProt[[1]][,1]),
               area2 = length(deProt[[2]][,1]),
               area3 = length(deProt[[3]][,1]),
               n12 = length(intersect(deProt[[1]][,1],deProt[[2]][,1])),
               n23 = length(intersect(deProt[[2]][,1],deProt[[3]][,1])),
               n13 = length(intersect(deProt[[1]][,1],deProt[[3]][,1])),
               n123 = length(Reduce(intersect, list(deProt[[1]][,1],deProt[[2]][,1],deProt[[3]][,1]))),
               category = c("a-c","l-c","o-c"),
               lty = "blank", fill = c("skyblue", "pink1", "mediumorchid"))
```
