==> 04-comparison.Rmd <==
#Comparison of the differents statistical test

Loading the proteins deemed as differentially expressed by the different statistical test.
```{r}
resSCL <- readRDS("resSCL.rds")
resLIMMA <- readRDS("resLIMMA.rds")
resANOVA <- readRDS("resANOVA.rds")
```


Venn diagram of the number of protein identified as differentialy expressed by the differents statistical test : 
```{r, message=TRUE}
library(VennDiagram)

grid.newpage()
draw.triple.venn(area1 = length(resSCL), 
                 area2 = length(resLIMMA), 
                 area3 = length(resANOVA), 
                 n12 = length(intersect(resSCL,resLIMMA)), 
                 n23 = length(intersect(resLIMMA,resANOVA)), 
                 n13 = length(intersect(resSCL,resANOVA)), 
                 n123 = length(Reduce(intersect, list(resSCL,resLIMMA,resANOVA))), 
                 category = c("scaling and centering", "LIMMA", "ANOVA"),
                 lty = "blank", fill = c("skyblue", "pink1", "mediumorchid"))
```

List of the proteins  identified as differentialy expressed only by ANOVA.
```{r}
setdiff(resANOVA,resSCL)
```

List of the proteins identified as differentialy expressed only by using LIMMA.
```{r}
setdiff(resLIMMA, resANOVA)
```
